Creation and validation of shader pipelines
The user can now form shader pipelines by combining shaders,
providing information about the pipeline input at the type level,
such as the memory layout of the vertex buffers in use,
and the primitive topology used for vertex assembly.
This allows the interfaces between consecutive shaders to be validated,
and information from this validation can be extracted for the creation
of a Vulkan graphics pipeline.

Many other changes, including:
  - manual specification of location/component information for vertex buffers,
  - memory layout now mostly computed at the type level,
    with custom type errors in case of invalid layouts,
  - some additional checks in place for accessing uniform buffers,
  - adding a logo, used as a window icon for the executable examples,
  - utility synonym for declaring standalone shaders
    (as opposed to defining entry points within a SPIR-V module),
  - improve type-inference of rebindable ifThenElse,
    by adding some constraints to the relevant typeclass instance context,
  - compilation of shaders now returns (Either Text ()).